requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord.py==1.4.1
idna==2.10
idna-ssl==1.1.0
multidict==4.7.6
numpy==1.19.1
pandas==1.1.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
typing-extensions==3.7.4.2
yarl==1.5.1
